Discord Bot2/requirements.txt
discord
wikipedia
pandas
youtube-search
pyowm
dhooks
googlesearch-python